Cut over fetchers to the streaming read path; remove fixed-window code (#266)
Fold src/meta_disco/streaming.py into fetchers.py as the production read path
(renaming fetch_*_streaming -> the canonical fetch_* names), and delete the old
fixed-window fetchers, the compare_streaming.py shadow-diff harness, and the
now-dead tests. The canonical fetcher names are reused, so file_types.py and
scripts/classify_hprc_files.py need no import changes.

- Deletes _read_head_until, _decompress_head/_decompress_if_gzipped,
  parse_tar_member_names, and the five old fetch_* bodies.
- Keeps the shared helpers streaming reuses (_fetch_range, _decode_bytes,
  extract_max_positions, parse_gfa_segment_tags, wrap_as_fetch_error,
  MAX_TAR_MEMBERS, TAR_HEAD_STAGES) and BAM (samtools) untouched.
- Merges the streaming unit tests into tests/test_fetchers.py; restores the
  tar member-cap warning and adds direct _fetch_range guard tests.

The evidence-cache clear and the corpus re-classification (VCF variants ->
variants.germline/somatic where the caller tag sits past the first BGZF block)
are the post-merge operational step, not part of this PR.
