+++ scripts/classify_tar_files.py
@@ -60,10 +60,10 @@ def main():
 
     if args.md5:
         file_name = args.filename or ""
-        # Default is_gzipped=True when the filename is unknown: _decompress_head only
-        # decompresses when the gzip magic bytes are present, so True is safe for a
-        # plain .tar too, whereas guessing False would skip decompression for a
-        # .tar.gz given by --md5 with no --filename (gzip bytes -> no members).
+        # Default is_gzipped=True when the filename is unknown: the streamed read only
+        # decompresses when the gzip magic bytes are actually present (_open_stream peeks
+        # them), so True is safe for a plain .tar too, whereas guessing False would skip
+        # decompression for a .tar.gz given by --md5 with no --filename (gzip bytes -> no members).
         result = classify_single_tar(
             args.md5,
             file_name=file_name,
